+++ lib/vespertine-std/src/fs/dir.rs
@@ -184,13 +184,13 @@ impl Dir {
     }
 
     pub fn subdir(&self, name: &'static str) -> Result<Dir, Error> {
-        let op = DirectoryOp::Lookup { name: name.as_ptr(), name_len: name.len() };
+        let op = DirectoryOp::Lookup { name: name.as_ptr() as usize, name_len: name.len() };
         let handle = sys_invoke(self.0, &Invocation::Directory(op)).map_err(Error::from)?;
         Ok(Dir::from(HandleID(handle)))
     }
 
     pub fn lookup(&self, name: &'static str) -> Result<HandleID, Error> {
-        let op = DirectoryOp::Lookup { name: name.as_ptr(), name_len: name.len() };
+        let op = DirectoryOp::Lookup { name: name.as_ptr() as usize, name_len: name.len() };
         let handle = sys_invoke(self.0, &Invocation::Directory(op)).map_err(Error::from)?;
         Ok(HandleID(handle))
     }
