+++ lib/vespertine-std/src/path.rs
@@ -5,7 +5,7 @@ use vespertine_rt::syscall::{SysError, sys_close, sys_invoke};
 pub fn walk_path(path: &str, root: HandleID) -> Result<HandleID, SysError> {
     let mut current = root;
     for segment in path.split('/').filter(|s| !s.is_empty()) {
-        let op = Invocation::Directory(DirectoryOp::Lookup { name: segment.as_ptr(), name_len: segment.len() });
+        let op = Invocation::Directory(DirectoryOp::Lookup { name: segment.as_ptr() as usize, name_len: segment.len() });
         let next = HandleID(sys_invoke(current, &op)?);
         if current != root { sys_close(current)?; }
         current = next;
